libraries/pyproject.toml
[tool.poetry]
name = "rw-public-codecollection"
version = "0.1.5"
description = "A set of RunWhen published keywords and python libs"
authors = ["Kyle Forster <kyle.forster@runwhen.com>"]
packages = [
  { include = "RW"}
]

[tool.poetry.dependencies]
python = "^3.8"
robotframework = "^4.1.2"
requests = "^2.26"
PyYAML = "^6"
boto3 = { version = "^1.21", optional = true }
awslimitchecker = { version = "^12", optional = true }
google-cloud-logging = { version = "^3", optional = true }
google-cloud-monitoring = { version = "^2.8", optional = true }
dnspython = { version = "^2.1.0", optional = true }
jira = { version = "^3.1.1", optional = true }
kubernetes = { version = "^19", optional = true }
opsgenie-sdk = { version = "^2.1", optional = true }
pdpyras = { version = "^4.3", optional = true }
pymsteams = { version = "^0.2", optional = true }
python-gitlab = { version = "^3.1", optional = true }
slack-sdk = { version = "^3.11", optional = true }
datadog = { version = "^0.44", optional = true }
"discord.py" = { version = "^1.7", optional = true }
hvac = { version = "^0.11", optional = true }
pygithub = { version = "^1.55", optional = true }


[tool.poetry.extras]
AWS = ["boto3", "awslimitchecker"]
GCP = ["google-cloud-logging", "google-cloud-monitoring"]
Hash = ["hvac"]
Datadog = ["datadog"]
Discord = ["discord.py"]
DNS = ["dnspython"]
Elasticsearch = []
GitHub = ["pygithub"]
GitLab = ["python-gitlab"]
Grafana = []
Jira = ["jira"]
K8s = ["kubernetes"]
MSTeams = ["pymsteams"]
Opsgenie = ["opsgenie-sdk"]
PagerDuty = ["pdpyras"]
Pingdom = []
Slack = ["slack-sdk"]


[[tool.poetry.source]]
name = "runwhen"
url = "https://backend-services.dev.project-468.com/pypi/platform"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
